API Expansions and Removals | Extracted editor logic from Flowchart
Expansions:
- Command has an overridable SkipExecution bool (returned Enabled by default) to make it clear when it's meant to serve a purpose other than executing logic
  - Label and Comment override it to always return true, for obvious reasons
- Added IVariableSource interface requiring a method for resetting the vars
- Added ICommandResetter interface
- Renames:
  - OnBlockStart -> BlockExecStart

Removals:
- BlockReference no longer holds a reference to a Block. Refactored the logic in it accordingly.
  - Much like VariableReference, it doesn't need to do its job through hard refs; storing config for
  fetching is what it's supposed to do
- Removed ColorCommands editor option
  - Why would you _not_ want commands colored?

Editor logic extracted from Flowchart:
- See the FlowchartEditorQol class
- FlowchartGlobalDefaults is the renamed FlowchartDefaultConfig.
  - The new name communicates the intent better.

Other:
- Renamed Block's command list to _legacyCommandList since eventually, we're
going to revamp Commands to be pocos
- Block and Command Ids are now bytes instead of ushorts

+++ Assets/Hyphlow/Runtime/Blocks/IBlockSource.cs
@@ -10,7 +10,7 @@ namespace AtMycelia.Hyphlow
         IReadOnlyList<IBlock> Blocks { get; }
         bool Contains(IBlock block);
         IBlock GetBlock(string name);
-        IBlock GetBlock(ushort id);
+        IBlock GetBlock(byte id);
     }
 
     /// <summary>
@@ -36,7 +36,7 @@ namespace AtMycelia.Hyphlow
         /// Returns true if the Block with the given id was successfully
         /// removed, false otherwise.
         /// </summary>
-        bool RemoveBlockWithId(ushort id, bool triggerSignals);
+        bool RemoveBlockWithId(byte id, bool triggerSignals);
         bool ClearBlocks(bool triggerSignals);
     }
 }
